user with authentication

+++ controllers/expensec.js
@@ -3,7 +3,7 @@ const Expense = require('../models/expense')
 module.exports.addExpensive= async (req,res,next)=>{
     try{
         const{amount,description,category}=req.body;
-        await Expense.create({amount, description, category })
+        await Expense.create({amount, description, category, userId : req.user.id })
         .then(()=>{
             res.status(200).json({message:'sucessfully added the expense'});
         })       
@@ -14,7 +14,7 @@ module.exports.addExpensive= async (req,res,next)=>{
 }
 module.exports.getExpensive = async(req,res)=>{
     try{
-        Expense.findAll().then(expense =>{
+        Expense.findAll({where:{ userId : req.user.id}}).then(expense =>{
             return res.status(200).json({ expense , success:true})
         })
         .catch(err =>{
@@ -31,7 +31,7 @@ module.exports.deleteExpensive = async(req,res)=>{
         if(expenseid == undefined || expenseid.length === 0){
             return res.status(400).json({success: false})
         }
-        Expense.destroy({where:{id: expenseid}}).then(()=>{
+        Expense.destroy({where:{id: expenseid, userId: req.user.id}}).then(()=>{
             return res.status(200).json({success: true, message: 'deleted successfully'})
         })
         .catch(err =>{
